+++ react/features/recording/components/LiveStream/native/StreamKeyForm.js
@@ -147,7 +147,7 @@ class StreamKeyForm extends AbstractStreamKeyForm<Props> {
      * @returns {void}
      */
     _onOpenGooglePrivacyPolicy() {
-        Linking.openURL(GOOGLE_PRIVACY_POLICY);
+        Linking.openURL(this.dataPrivacyURL);
     }
 
     _onOpenHelp: () => void
@@ -176,7 +176,7 @@ class StreamKeyForm extends AbstractStreamKeyForm<Props> {
      * @returns {void}
      */
     _onOpenYoutubeTerms() {
-        Linking.openURL(YOUTUBE_TERMS_URL);
+        Linking.openURL(this.termsURL);
     }
 }
